[DM] Device mapper core framework, implement linear/striped/zero/verity/error targets, ioctl handlers and LVM test scripts

+++ kernel/core/comps/block/src/partition.rs
@@ -251,7 +251,7 @@ pub struct PartitionNode {
 
 impl BlockDevice for PartitionNode {
     fn enqueue(&self, mut bio: SubmittedBio) -> Result<(), BioEnqueueError> {
-        bio.set_sid_offset(self.info.start_sector());
+        bio.set_sid_offset(self.info.start_sector() as i64);
         self.device.enqueue(bio)
     }
